Add origin version.

+++ Assets/GameMain/Scripts/Network/PacketHeaderBase.cs
@@ -17,11 +17,13 @@ namespace StarForce
             get;
         }
 
-        public int Id
-        {
-            get;
-            set;
-        }
+        //public int Id
+        //{
+        //    get;
+        //    set;
+        //}
+
+        public ushort Id { get; set; }
 
         public int PacketLength
         {
